refactor(firebase): move files into more sub-folders
