+++ src/features/firebase/types/guilds/channels/features/firebase-guild-channel-feature.ts
@@ -1,0 +1,3 @@
+import { IFirebaseGuildChannelFeatureV1 } from "../../../../interfaces/guilds/channels/features/firebase-guild-channel-feature-v1";
+
+export type IFirebaseGuildChannelFeature = IFirebaseGuildChannelFeatureV1;
