+++ src/features/firebase/types/guilds/channels/firebase-guild-channel-v-final.ts
@@ -1,0 +1,3 @@
+import { IFirebaseGuildChannelV1 } from "../../../interfaces/guilds/channels/firebase-guild-channel-v1";
+
+export type IFirebaseGuildChannelVFinal = IFirebaseGuildChannelV1;
