--- src/features/firebase/types/guilds/firebase-guild-channel-feature.ts
@@ -1,3 +1,0 @@
-import { IFirebaseGuildChannelFeatureV1 } from "../../interfaces/guilds/firebase-guild-channel-feature-v1";
-
-export type IFirebaseGuildChannelFeature = IFirebaseGuildChannelFeatureV1;
