--- src/features/firebase/types/guilds/firebase-guild-channel-v-final.ts
@@ -1,3 +1,0 @@
-import { IFirebaseGuildChannelV1 } from "../../interfaces/guilds/firebase-guild-channel-v1";
-
-export type IFirebaseGuildChannelVFinal = IFirebaseGuildChannelV1;
