--- src/features/firebase/types/guilds/firebase-guild-channel.ts
@@ -1,3 +1,0 @@
-import { IFirebaseGuildChannelV1 } from "../../interfaces/guilds/firebase-guild-channel-v1";
-
-export type IFirebaseGuildChannel = IFirebaseGuildChannelV1;
